requirements.txt
dj-database-url==0.5.0
Django==3.0.3
django-allauth==0.44.0
django-cors-headers==3.2.1
django-filter==2.3.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
Markdown==3.3.3
oauthlib==3.1.0
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.14.0
python-xlib==0.27
python3-openid==3.2.0
requests==2.23.0
requests-oauthlib==1.3.0

